+++ src/src/components/celebrity/CelebInfoGridCard.jsx
@@ -108,7 +108,7 @@ function CelebInfoGridCard({
           alt={`${celebName} 프로필 사진`}
         />
       ) : (
-        <TestAccountIcon size={100} />
+        <TestAccountIcon size={"100"} />
       )}
 
       <Styled.TextContainer>
