+++ src/src/components/common/button/TabButtons.jsx
@@ -41,13 +41,13 @@ const Styled = {
 };
 
 /**
- * Tabs 공통 컴포넌트
+ * TabButtons 공통 컴포넌트
  * @param {Array<{title: string, func: function}>}tabInfoArray title과 func 정보를 객체로 담은 정보 배열
  * @param {Object} tabStyle 기존 style과 동일
  * @param {React.htmlAttributes} tabBtnContainerProps 기타 탭 버튼 container props
  */
 
-function Tabs({ tabInfoArray, tabStyle, ...tabBtnContainerProps }) {
+function TabButtons({ tabInfoArray, tabStyle, ...tabBtnContainerProps }) {
   const [selectedTab, setSelectedTab] = useState(0);
 
   return (
@@ -55,7 +55,7 @@ function Tabs({ tabInfoArray, tabStyle, ...tabBtnContainerProps }) {
       {tabInfoArray.map((tab, index) => (
         <Styled.Tab
           value={index}
-          onClick={(event) => {
+          onClick={event => {
             if (selectedTab === event.target.value) return;
             setSelectedTab(event.target.value);
             tab.func();
@@ -71,11 +71,11 @@ function Tabs({ tabInfoArray, tabStyle, ...tabBtnContainerProps }) {
   );
 }
 
-Tabs.propTypes = {
+TabButtons.propTypes = {
   tabInfoArray: PropTypes.arrayOf(
     PropTypes.shape({ title: PropTypes.string, func: PropTypes.func }),
   ).isRequired,
   tabStyle: PropTypes.object,
 };
 
-export default memo(Tabs);
+export default memo(TabButtons);
